test(parser): split tests in multiple files
